Rename package in app module
