+++ app/src/main/java/com/yesferal/hornsapp/app/MainActivity.kt
@@ -1,8 +1,7 @@
-package com.yesferal.hornsapp
+package com.yesferal.hornsapp.app
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : AppCompatActivity() {
